+++ app/src/main/java/com/booboot/vndbandroid/model/vndb/AccountItem.kt
@@ -1,6 +1,6 @@
 package com.booboot.vndbandroid.model.vndb
 
 abstract class AccountItem {
-    abstract var vn: Int
+    abstract var vn: Long
     abstract var added: Int
 }
