+++ app/src/main/java/com/booboot/vndbandroid/repository/TagsRepository.kt
@@ -25,10 +25,10 @@ class TagsRepository @Inject constructor(
     var db: DB,
     var boxStore: BoxStore
 ) : Repository<Tag>() {
-    override fun getItems(cachePolicy: CachePolicy<Map<Int, Tag>>): Single<Map<Int, Tag>> = Single.fromCallable {
+    override fun getItems(cachePolicy: CachePolicy<Map<Long, Tag>>): Single<Map<Long, Tag>> = Single.fromCallable {
         cachePolicy
             .fetchFromMemory { items }
-            .fetchFromDatabase { /* TODO */ boxStore.boxFor<TagDao>().all.map { it.toBo() }.associateBy { it.id } }
+            .fetchFromDatabase { boxStore.boxFor<TagDao>().all.map { it.toBo() }.associateBy { it.id } }
             .fetchFromNetwork {
                 vndbService
                     .getTags()
@@ -45,7 +45,6 @@ class TagsRepository @Inject constructor(
             }
             .putInMemory { items = it.toMutableMap() }
             .putInDatabase {
-                /* TODO */
                 with(boxStore.boxFor<TagDao>()) {
                     put(it.map { TagDao(it.value, this) })
                 }
