+++ app/src/main/java/com/booboot/vndbandroid/ui/vndetails/VNDetailsTabsAdapter.kt
@@ -23,7 +23,7 @@ class VNDetailsTabsAdapter(fm: FragmentManager?) : FragmentStatePagerAdapter(fm)
 
     override fun getItem(position: Int): Fragment {
         val args = Bundle()
-        args.putInt(VNDetailsActivity.EXTRA_VN_ID, vn?.id ?: 0)
+        args.putLong(VNDetailsActivity.EXTRA_VN_ID, vn?.id ?: 0)
 
         val fragment = when (position) {
             0 -> TagsFragment()
