+++ app/src/main/java/com/booboot/vndbandroid/ui/vndetails/VNDetailsViewModel.kt
@@ -18,7 +18,7 @@ class VNDetailsViewModel constructor(application: Application) : BaseViewModel(a
         (application as App).appComponent.inject(this)
     }
 
-    fun loadVn(vnId: Int, force: Boolean = true) {
+    fun loadVn(vnId: Long, force: Boolean = true) {
         if (!force && vnData.value != null) return
         if (disposables.contains(DISPOSABLE_VN)) return
